fix names and show product-preview
